Move GF functions into WP folder
