+++ components/molecules/GravityForm/Fields/Checkbox.js
@@ -1,7 +1,10 @@
-import PropTypes from 'prop-types'
-import {getGfFieldId, getGfHiddenClassName} from '@/functions/gravityForms'
-import cn from 'classnames'
 import * as Input from '@/components/atoms/Inputs'
+import {
+  getGfFieldId,
+  getGfHiddenClassName
+} from '@/functions/wordpress/gravityForms'
+import cn from 'classnames'
+import PropTypes from 'prop-types'
 
 /**
  * Merge GravityForm checkbox field data into single Array.
